updated the api create param to include cr objects (#5)
* updated the api create param to include cr objects

* fixed pool cr status field

* disabled pool cr deletion

* fix cr status race condition

* fixed memory negativity

* fixed memory negativity

+++ api/v1alpha1/storagenode_types.go
@@ -72,7 +72,7 @@ type NodeStatus struct {
 	Health    string        `json:"health,omitempty"`
 	Status    string        `json:"status,omitempty"`
 	CPU       *int32        `json:"cpu,omitempty"`
-	Memory    *int32        `json:"memory,omitempty"`
+	Memory    *int64        `json:"memory,omitempty"`
 	Volumes   *int32        `json:"volumes,omitempty"`
 	RPC_PORT  *int32        `json:"rpc_port,omitempty"`
 	LVOL_PORT *int32        `json:"lvol_port,omitempty"`
